requirements.txt
appnope==0.1.3
asttokens==2.2.1
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
customtkinter==5.0.3
darkdetect==0.7.1
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
future==0.18.2
geocoder==1.38.1
h11==0.14.0
idna==3.4
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
outcome==1.2.0
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pyzmq==24.0.1
ratelim==0.1.6
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
tkintermapview==1.18
tornado==6.2
tqdm==4.64.1
traitlets==5.7.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wcwidth==0.2.5
webdriver-manager==3.8.5
wsproto==1.2.0
